Documenting the errors CategoryController can throw

+++ src/category/category.controller.ts
@@ -1,5 +1,12 @@
 import { InvalidIdException } from "@/shared/exceptions";
 import {
+  BadRequestResponse,
+  InternalServerErrorResponse,
+  NotFoundResponse,
+  ServiceUnavailableResponse,
+} from "@/shared/swagger";
+import { InvalidDataException } from "@/shared/validation/exceptions";
+import {
   Controller,
   Get,
   Post,
@@ -8,7 +15,12 @@ import {
   Param,
   Delete,
 } from "@nestjs/common";
-import { ApiTags } from "@nestjs/swagger";
+import {
+  ApiBadRequestResponse,
+  ApiInternalServerErrorResponse,
+  ApiNotFoundResponse,
+  ApiTags,
+} from "@nestjs/swagger";
 import { omitBy } from "lodash";
 import { CategoryService } from "./category.service";
 import { CreateCategoryDto, UpdateCategoryDto } from "./dto";
@@ -19,16 +31,24 @@ export class CategoryController {
   constructor(private readonly categoryService: CategoryService) {}
 
   @Post()
+  @BadRequestResponse(InvalidDataException)
+  @InternalServerErrorResponse()
+  @ServiceUnavailableResponse()
   public async create(@Body() createCategoryDto: CreateCategoryDto) {
     return this.categoryService.create(createCategoryDto);
   }
 
   @Get()
+  @InternalServerErrorResponse()
+  @ServiceUnavailableResponse()
   public async findAll() {
     return this.categoryService.findAll();
   }
 
   @Get(":id")
+  @BadRequestResponse(InvalidIdException)
+  @InternalServerErrorResponse()
+  @ServiceUnavailableResponse()
   public async findOne(@Param("id") id: string) {
     if (!+id) {
       throw new InvalidIdException();
@@ -37,6 +57,10 @@ export class CategoryController {
   }
 
   @Patch(":id")
+  @NotFoundResponse()
+  @BadRequestResponse(InvalidIdException, InvalidDataException)
+  @InternalServerErrorResponse()
+  @ServiceUnavailableResponse()
   public async update(
     @Param("id") id: string,
     @Body() updateCategoryDto: UpdateCategoryDto,
@@ -49,7 +73,13 @@ export class CategoryController {
   }
 
   @Delete(":id")
+  @BadRequestResponse(InvalidIdException)
+  @InternalServerErrorResponse()
+  @ServiceUnavailableResponse()
   public async remove(@Param("id") id: string) {
+    if (!+id) {
+      throw new InvalidIdException();
+    }
     return this.categoryService.remove(+id);
   }
 }
